requirements.txt
setuptools==67.8.0
pytest==7.3.1
selenium==4.9.1
jproperties==2.1.1
paramiko~=3.2.0
names~=0.3.0
Faker==18.10.1
shortuuid==1.0.11
requests~=2.31.0
python-dateutil~=2.8.2
allure-python-commons==2.13.2
